requirements.txt
fastapi
pycryptodome
pillow
